// src/app/masters/models/company.ts
export class Company {

    companyname: any;
    shortname: any;
    address: any;
    mobile: any;
    phone: any;
    fax: any;
    email: any;
    web: any;
    district: any;
    postcode: any;
    ircno: any;
    createdby: any;
    createddate: any;
    imagepath : any;
    address2: any;
    deleted: any;
    lcano: any;
    yearofrenewal: any;
    tinno: any;
    vatregno: any;
    ercno: any;
    hscode: any;
    sectorofindustry: any;
    constructor() {}
}